fix: refine glass dropdown motion
