+++ src/components/ui/index.ts
@@ -2,8 +2,7 @@ export { AnimatedTabs } from "./AnimatedTabs";
 export type { AnimatedTabItem } from "./AnimatedTabs";
 export { Button } from "./Button";
 export { ConfirmModal } from "./ConfirmModal";
-export { Dropdown } from "./Dropdown";
-export { DropdownItem } from "./DropdownItem";
+export { Dropdown, DropdownItem } from "./Dropdown";
 export { EmptyState } from "./EmptyState";
 export { IconButton } from "./IconButton";
 export { Modal } from "./Modal";
